requirements.txt
# Core ML libraries
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0

# Game interface
selenium>=4.15.0
playwright>=1.40.0

# Utilities
pyyaml>=6.0
wandb>=0.16.0
tqdm>=4.66.0
matplotlib>=3.8.0
tensorboard>=2.15.0

# Development
pytest>=7.4.0
black>=23.12.0

